Update DFS.java
had your runMapReduce to implement a thread in itself so you can have the thread sleep a bit before it try to check and see if peer has anything in the set. Without the sleep it will check the peer's set instantaneous without having the peer populate anything into it's set.