fixed calling mapContext
The way mapContext should work is that any node can tell a node to do their work. 

Lets say there are 4 nodes, and node #2 has the metadata
If you use node 3 to do mapreduce, you want it to be able to talk to node #2 and tell it to do their work.
You might run into some error with this code previously because hypothetically node #2 has the metadata and the files, but you are using node #3 to try and get the file form his own repository, same concept is applied to reduce, but with a twist.